criei uma nova pasta chamada controllers para desmembrar melhor o código assim permitindo melhor manutençao
